requirements.txt
amqp==5.0.6
asgiref==3.4.1
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
djangorestframework==3.12.4
flower==0.9.2
kombu==5.2.1
prompt-toolkit==3.0.22
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
